Modify FOV method to return double instead of float
Changed the return type of the silvera_modifyFov method from float to double to accommodate changes in FOV calculations.

+++ src/main/java/exloran/silvera/mixin/GameRendererMixin.java
@@ -1,26 +1,20 @@
 package com.silvera.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import com.silvera.module.impl.cosmetic.ZoomModule;
-import com.silvera.render.HudRenderer;
-import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.render.GameRenderer;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyReturnValue;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 /**
  * GameRendererMixin
- *  - FOV modifikasyonu (Zoom)
- *  - HUD render hook
+ * - Zoom için FOV değiştirir
  */
 @Mixin(GameRenderer.class)
 public class GameRendererMixin {
 
-    /** Zoom: getFov'un dönüş değerini yakala ve modifiye et */
     @ModifyReturnValue(method = "getFov", at = @At("RETURN"))
-    private float silvera_modifyFov(float original) {
+    private double silvera_modifyFov(double original) {
         return ZoomModule.getModifiedFov(original);
     }
 }
